Added Get functions to the new API V3: Download Eth0; Download Wlan0; Upload Eth0; Upload Wlan0.
Refactor on the Controller and get_func packages.
