+++ common/url_constructor.py
@@ -30,7 +30,9 @@ class URLs:
             'statusWireless': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v1/interface/wireless/1/status',
             'statusSystem': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v1/system/device/status',
             'HasUpdate': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v1/system/apply/status',
-            'WanInfo': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v1/interface/wan'
+            'WanInfo': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v1/interface/wan',
+            'throughputEth0': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v1/service/statistics/wlan0',
+            'throughputWlan0': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v1/service/statistics/wlan0'
 
         }
         return URL[self.info]
@@ -44,11 +46,14 @@ class URLs:
             'login': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v3/system/login',
             'clients': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v3/system/status',
             'version': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v3/system/device',
-            'noise': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v3/interface/wireless/1/survey',
-            'statusWireless': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v3/interface/wireless/1/status',
+            'noise2Ghz': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v3/interface/wireless/radio0/survey',
+            'noise5Ghz': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v3/interface/wireless/radio1/survey',
             'statusSystem': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v3/system/device/status',
             'HasUpdate': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v3/system/apply/status',
-            'WanInfo': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v3/interface/wan'
+            'WanInfo': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v3/interface/wan',
+            'throughputEth0': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v3/service/statistics/eth0',
+            'throughputWlan0': 'https://' + self.ip + ':' + self.port + '/cgi-bin/api/v3/service/statistics/wlan0'
+
 
         }
         return URL[self.info]
